refactor(ui): finalize v1.5 design system rail-anchored migration
